% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{createIFS}
\alias{createIFS}
\title{IFS S3 constructor}
\usage{
This object can then be plotted, see: plot.IFS
}
\arguments{
\item{X}{Vector or list of functions of type function(twoDimVector){ ...; return( c(x1, y1) ); } (which we hope are also contractions on R^2)}
}
\value{
IFS object or NULL if argument doesn't consist of functions or some other error occured
}
\description{
This creates a new IFS object that we can plot
}
\examples{
createIFS(IFSkoolnamr::sierpinski)
}
