% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{dragon}
\alias{dragon}
\title{Dragon-like curve}
\format{An object of class \code{list} of length 2.}
\usage{
Recommended k > 15
}
\description{
Dragon-like curve
}
\keyword{datasets}
