% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ifs.R
\name{plot.IFS}
\alias{plot.IFS}
\title{Plot IFS object}
\usage{
plot.IFS(.obj, k = 7, RES = c(500, 500), KOL = "red")
}
\arguments{
\item{k}{iterations of contractions we will perform}

\item{RES}{resolution of pixels drawn onto a plot (points go from floats to integers at some point)}

\item{KOL}{colour of the image, e.g. "red", "green", "black", "#20592b".}
}
\value{
Raw data used for plotting and a plot as a by-product
}
\description{
Plot each function given in constructor with k iterations
}
\examples{
plot(createIFS(IFSkoolnamr::sierpinski), k = 8)
}
