% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{sierpinski}
\alias{sierpinski}
\title{Sierpinski's triangle}
\format{An object of class \code{list} of length 3.}
\usage{
Recommended k > 10
}
\description{
Sierpinski's triangle
}
\keyword{datasets}
